Change how the testing channel is implemented
