+++ .github/workflows/build-deploy.yml
@@ -54,6 +54,7 @@ jobs:
             secret/data/keybase/bots/githubbot webhook_receiver | KEYBASE_WEBHOOK_URL;
             secret/data/alertmanager-receiver/okta-app client-secret | WEBHOOK_AUTH_SECRET_TOKEN;
             secret/data/keybase/bots/githubbot webhook_testing_channel | TESTING_KEYBASE_WEBHOOK_URL;
+            
       - name: Login to k8s cluster
         uses: Chia-Network/actions/vault/k8s-login@main
         with:
