+++ cmd/notifyStale.go
@@ -26,6 +26,7 @@ var notifyStaleCmd = &cobra.Command{
 			slogs.Logr.Fatal("Error loading config", "error", err)
 		}
 		client := github.NewClient(nil).WithAuthToken(cfg.GithubToken)
+		webhookURL := "https://alert-receiver.chiaops.com/devrel"
 
 		datastore, err := database.NewDatastore(
 			viper.GetString("db-host"),
@@ -92,7 +93,7 @@ var notifyStaleCmd = &cobra.Command{
 					description := pr.URL
 					slogs.Logr.Info("Sending message via keybase for", "repository", pr.Repo, "PR", int64(pr.PRNumber))
 					message := keybase.NewMessage(status, title, description)
-					if err := message.SendKeybaseMsg(); err != nil {
+					if err := message.SendKeybaseMsg(webhookURL); err != nil {
 						slogs.Logr.Error("Failed to send message", "error", err)
 						time.Sleep(15 * time.Second) // This is to prevent "error response: 429 Too Many Requests""
 					} else {
